refactor: nest sharedData under dummyData
