+++ src/dummyData/sharedData/styles.ts
